requirements.txt
Django==1.11.20
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.1
Pillow==5.4.1
pytz==2018.9
